fix(DroHub): Improved user creation authorization chain
Specifically it is now defined by code and test that the create user functionality
can only create a new Subscription or add a new user. It cannot edit it's values
like time allowed or user count allowed. That functionality needs to be implemented,
and for now the only way to achieve it is by deleting the last user of the Subscription.

In the AdminPanel.cs we changed from condition checks to methods that will throw if
their conditions are not met. This allows for more concise code in the OnPostAsync.

Check if the current subscription allows for new users, if the subscription already
exists. This is so that even an admin cannot violate the current subscription limits,
as editing subscriptions on the user create is disallowed.

There were test additions to test that this is true.