feat(DroHub): Added support for time limited subscription
In this commit changes are made so that the subscription allowed time is
taken into account and monitored during a flight session in the
DroneMicroServiceManager.

The IThriftTask was changed so that it has a
precondition callback that is called before the socket is accepted. This
is so that we keep the subscription checking logic all in the
DroneMicroServiceManager. If it fails we just set bad request and
shortcircuit the middleware.

TestThriftDrone test was overhauled so
that it can test that the allowed time is enforced. This test can still
be improved and does not yet check the case of multiple concurrent
drones. Importantly it implements an idea of checking connection health
which should be implemented in the drone application: The concept that
if no data was requested in more than a given amount of time, then we
should consider the connection dead on the client side and act
accordingly.