test(DroneDeviceHelper): Add a cancellation token to the TelemetryItem
In the WaitForServer besides the telemetry item tasks
we had an extra task with a timeout. The task would be a
canary for when it took too long to retrieve the telemetry
item from the remote. This system did not provide much
information about what telemetry was not completed
successfully and provided a single cut off for all
telemetry systems. I also think it did not work.

Now we have a cancellation token with a callback that
cancels the task source when timeout occurs. This is
per telemetry item allowing discriminate failure reporting.