refactor(ConnectionManager and ThriftMessageHandler)
Move a lot of the cleanup in ThriftMessageHandler into the
Dispose method.

ConnectionManager now is not much more than a glorified dictionary
with helper methods. This is to stop any temptation of manipulating
ThriftMessageHandlers from the ConnectionManager. To the point that
ConnectionManager no longer tries to manage pre-existing connections.
Instead when ThriftMessageHandler gets a connection and is about to
add it to the ConnectionManager, it check whether a connection for
the same device already exists. If it does, it cancels the previous
connection, removes it from the ConnectionManager and only after a
successful connection on the websocket does it add it to the
connection manager again.

ConnectionManager also now stores the time at which the connection
was started so that users can retrieve it by GetConnectionStartTime.
No users are using this at this time so no functionality was affected