Version(0.1) Major refactor and features internal to DroHub web.
All subscription operations now go through an API service called SubscriptionAPI.cs

Authentication of the AndroidApplicationController is now done with a special
authentication scheme that verifies that special HTTP headers for the user
and the token exist. This way credentials are no longer passed by request body
but by headers. Added a DroHubObjectRequest.java class that can be used to
prepare the requests to DroHub web. For the thrift middleware we actually use
a more thorough authentication scheme which checks that the device is included
in the headers and is valid.

Removed AuthenticationToken action from AndroidApplicationController because
it was redundant, and did nothing besides saying the authentication was valid.
There was no scenario where this was useful currently. Android application was
adapted.

Added SubscriptionOrganizationName to Device model, so that we can directly expose
the foreign key of the device subscription. This is really helpful so we do not
need to include the full subscription when we want to know the device's
subscription.

Moved claim names from the Device model to a new resource authorization handler
for Devices called DeviceAuthorizationHandler. This handler should be used by
the DeviceAPI to authorize operations on Device resources. Added a base class
with the basic CRUD operations on resources so that Resource authorization handlers
can extend or use.

Renamed DeviceHelper to a new Service API for Device operations.
All device related operations including querying, manipulation and authorization
should be done through this transient service. It also includes a struct
DeviceSerial which wraps the device serial number when passed around. This
allows for type safety of the serial number type. It is a bit non ideal as
a simple typedef or type aliasing would have been enough :(.

In this version the initial values are not included in the devices for the
DeviceRepositoryController.

Fixed a long bug where the GetTelemetry family of actions in DeviceController
was returning a List with a list. This was discovered when moving telemetry
functionality to DeviceAPI.cs

Almost complete re-write of DroneMicroService to become
DroneMicroServiceManager.cs.
This to avoid race conditions in database usage as well as other shared resources.
Maybe in the future each of the services will be factored out, but for now they
are so small individually that its ok to keep them all in the same place.

Removed EchoStream.cs and re-did the websocket to thrift bridge with MemoryStreams
and a new Stream class called ChannelInputStream. ChannelInputStream takes
a channel of byte arrays and consumes it when it is read. After a successful read
operation the undelying memory stream is refreshed so that it does not grow
infinitely. This channel scheme was the answer to massive thread exhaustion
issues, where threads were blocked waiting for data to be fed. The maths was
that for each connection there were 6 telemetry feeds concurrently with 2
EchoStream each. When the number of connection went to 30 or 40 we had hundreds
of threads blocked and IO started to timeout on random places like the Database
or the Janus operations. Currently we are able to test 100 concurrent connections
at the same time with no issue. The profiler showed that the EchoStream took
60% of the time of execution for a single connection.

Separated GetTokenController out of AndroidApplicationController so that we
could have a light weight controller that did not require authenticated services.

Added subscription organization name to avoid database round trips. This is fine
because it is not expected that the subscription of the current user changes mid
session.

JanusService now catches OperationCanceled thrown by the HttpClient, because they
are not cancellations from our application but due to timeout in the http
request it does. We now rethrow it as a JanusServiceException so that we do
not hide the exceptions happening on HttpClient, as they are fatal to the live
video micro service. Increased the timeout to 60 seconds just in case.

Automatically updated JDK version to 11.

Added a RouteAuthenticationHandler that just routes the authentication schemes
to be used according to the paths of the HttpContext request. Very useful for
different authentication requirements in different end points.

Minor refactor and additional logging and checks on TAMessageValidatorProtocol. Now
we validate not only the id but also the message name. A bit redundant but that's
it. Changed from if-else to switch statements.
Deleted py interfaces.

We no longer support a python client so lets
remove obsolete stuff.