fix(DroHub): Fix dependency injection failing on DeviceAuthorizationHandler
For a weird reason the DeviceAuthorizationHandler is constructed before
authentication is run, leading to the SubscriptionAPI failing to
construct. We hack around by only instantiating the SubscriptionAPI upon
authorization. As the authorization is the only public methods it is
fine.

The above error was only triggered when a user tried to access a url which he was
not authorized for. The normal case was fine. So we also add a new test case
that verifies that we are actually redirected to the login page when we try to
access a page that requires authentication. A small adaptation to HttpClientHelper
was required because createHttpClient did not check for the status code.