feat(DroHub front and back): Added support for flight time
Now the video overlay shows the flight time as measured since the
application connected to DroHub.

Added a dummy test that does just simulates a very long static flight
session. Probably should animate it in the future. Used it to test
the visual part of the overlay updating the flight time. To accomodate
not needing to create a user and use the default admin user some
changes were made in the drone simulation code.

On the dashboard.js a TelemetryWatchdogClass was added that triggers
some functions on start or timeout. A function that refreshes non
drone telemetry data also was created. This was needed because the
telemetry of a flight time is a visual event not really a drone event.

Updated signalR javascript file to reflect the 3.1 version.