feat(Sidebar) Improved visuals and added acting role
More work toward finalizing the the sidebar visually and now with an
added Role below the user name. For this we needed to add a Claims
Identity Factory that populates the claim with the acting role. This is
required so we do not always query the database, with basic info from the
user.