feat(codec); Use VP8
Let's try to use VP8 which seems to be an actual web standard and
not use H264 which seems to be very complicated and have a less performant
encoder in android devices. This is part of some live experiments to
figure out what gives the best experience.

Also it seems that we should not cap the bitrate and fps and just let
the webrtc stack handle it. Limiting the FPS specifically leads to very
poor stuttering. bitrate limiting also.

Limited the video resolution, because the ANAFI live video camera is
actually 720p not the whole screen size.