feat(Gallery); Populate sidebar server side
We did not have any dynamic reloading of the devices available, yet
we did the population of the sidebar device list in the client side
at the cost of an extra connection. Due to the fact we now have a
DeviceAPI service we can just inject it and use it to get the subscription
devices. Very happy with this simplification.