feat(Gallery); Moved Gallery to DeviceRepositoryController
The DevicesController should only concern itself with single Device
operations, while the DeviceRepositoryController concerns itself
with collections of devices. This is the first commit that will
turn the gallery into a multimedia dashboard for all the subscription's
devices instead of the individual gallery for each device we currently
have. The reason i split this commit is so that it's more manageable.