feat(janus) Add support for the admin console.
The admin console allows to get extra diagnostics about the video
rooms and respective peers that otherwise we would not easily
have access to. The admin console is only accessible through the
127.0.0.1 address meaning only local connections are able to
connect to it. The secret password of the admin console is version
tracked but should be one day be an environmental variable passed
by the docker compose file