feat(janus.jcfg) Improve janus defaults.
We now have a limited rtp_port_range so that we do not need to
open all the up port range. Firewall settings should match the
range or some sessions will randomly not work.

Reverted to use the google stun server instead of xirsys because
we are now not paying xirsys and they have limited amount of stun
requests per month.

Also removed the turn server settings because with the current cloud
deployment a simple stun request and the hardcoding of the public IP is
enough.