feat(janus,Drohub,app) Update and improvement.
Updated the Janus Dockerfile with ubuntu 20.04 which is an LTS. This
means that janus now is 0.7.3. Some regressions happened, namely
firefox video no longer works and the videoroomtest.js needed to be
changed to always publish the audio even if this was not in the plans
yet. Still in videoroomtest.js added some better debugging printouts.

Needed to add a default videoDecoderFactory because otherwise with the
new janus the Android app just crashes.

Also added vim and janus demos packages so that some debugging can
be done with the demo rooms.

Removed unused janus configurations and strictly added the configurations
we require. This includes adding the videoroom plugin configuration
and disable the demo room, which previously was enabled by default.