feat(Gallery) Single gallery for all devices.
We now have the gallery grouped by date and then by device, with the
videos being tiled in such subdivision. Only basic styling was applied
as the goal was to get rid of the gallery per device. Also there is
a small bug where the movies are grouped by UTC day not user locale
in the server side. This is because the user locale is not know in
the server side. It is worth to mention though that the dates are shown
according to the user locale.