fix(app) Show waiting screen on application when no device is connected.
There was a regression where when a device was not connected, we would
never try to authenticate. This led to strange behavior where the login
would always be shown even though we already had a valid token. This
regression happened when we removed the validateToken action and were
left with no way to validate the token without a device serial. We work
around this by using a NODEVICE serial number when we do not have a
device serial number but use the queryDeviceInfo flow to get an empty
answer but with no errors if the token is valid. Added logic to disallow
the NODEVICE serial number to never be allowed for registration.