fix(JanusService) Added FirFrequency parameter.
This parameter makes janus request publishers for an inter frame
at a given frequency. In this case we set it to every 7 seconds. This
is useful to avoid corruption of frames and long intervals where the
intraframes are sent with not interframe as a reference, leading to
unusable movies. This probably should become a configuration setting
but for now this is a hardcode.