feat(MJR-Converter) Added ability to also parse audio mjr. Now with tests
This review has major refactor, almost a rewrite of MJR converter. The
refactor was so that we could easily add the ability to generate other
media files from MJR files generated by Janus. For example we now read
the file extension of the post process of mjr files from the MJR header.

Also added a test of MJR-Converter as well as related assets.
MJR-Converter is now part of the solution also.