feat(DroHub): Added support for DeviceConnection data.
We now have a persistent data structure that tracks all
the device data from a given connection, including it's
start and end. It is not very useful at this moment but
this structure will allow for very granular session data.

Mostly all the telemetry went from being associated with a
device to becoming associated to a DeviceConnection, whether
ongoing or past one. This simple change led to very deep changes
in the backend, and not everything is tested at this moment
as the features that will use this infrastructure are not
yet developed.

This is a commit with a lot of secondary improvements but
unfortunately untested ones.