feat(DroHub): MJR conversion is now done on request of GatherDeviceLiveVideoService.
In this commit we dismantle the separate MJR converter and make it a helper of
DroHub. This is required because we want to stop relying on recording file name
and want to be able to relate media to a device connection session. We achieve
this by making the creation of the room set up a recording directory based on the
connection id. This way we can know inside Drohub all the media that a session
produced in a given flight session, by looking at the connection session directory.

Due to the grouping mentioned we now are able to group not only video but also audio
in the gallery.

There are quite some testing regressions on this commit. We no longer have a test for
the mjr converter, and will probably need to create a unit test for it.

When testing locally, if drohub is running outside docker it will not have permissions
to post process the mjr files.