feat(DroHub): Huge amount of changes for the gallery
_Main.cshtml moved scripts section to head section of the html, as it
allows consistency of loading with other assets. It also allows for
dynamic pages to write on the body of the html document without fear
of losing the scripts and assets loaded in the head of the document.

_Main.cshtml Moved icheckbox initialization out of the _Main.cshtml
and put it in scripts where it is needed. This allows for customization
per page.

_Main.cshtml: Added a new razor pages section for the head so we can
add assets to the head, for example on a partial view.

isFunction and addFunction were moved to site.js as these functions
are useful in other places besides the dashboard.js.

DeviceConnection.cs now has a new list of MediaObject class associated.
This way we can search for media objects of a given connection.

DeviceConnectionAPI.cs: DeviceConnectionException constructor is now
internal so that it cannot be instantiated outside the assembly.

DeviceConnectionAPI.cs: Connection media dir and related information
was moved into the MediaObjectAndTagAPI.

getSubscribedDeviceConnection is now pat of DeviceConnectionAPI and
includes a fat database set because we need it for the gallery. It
probably can be broken down later on but now it is useful. This move
from the SubscriptionAPI made the need of a SubscriptionAPI object in
DeviceRepositoryController.cs redundant.

DeviceRepositoryController has a re-work on the model passed to the
gallery view, for more rich content, although the organization and
concept did not change.

DeviceRepositoryController now has a deleteMediaObject method to
delete a media object from the gallery. Also has a getFile that
allows for the download of the object. Currently the file needs
to be a video. No tests are prepared for these functionalities
unfortunately. Add and delete tag functionality also added.

Added MediaObjectTag and MediaObject models. They provide us a way
to track information on the media, creation time and connection
session associations. Added a respective MediaObjectAndTagAPI to
let clients manage these objects in a controlled way. Also an
Authorization handler for the MediaObject resource was created
to manage the permissions on the assets. Associated claims were
added to the DroHubUser.

Overhauled the MJR to webm file so as to combine the streams of
audio and video into one file through ffmpeg. This system is now
a bit stupid on the syncronization and needs re-work, but the concept
and infrastructure for tweaking ffmpeg command is in place.

Also added a custom video-player overlay that controls video playback.
So far it does it's work although some details are still rough, namely
it needs to not fight with the user when the video is in playback and
the user tries to scroll. It also requires a full screen button.

The gallery front end now allows for selection of tags, as well as
editing tags in place.

Changed some svg assets so that we can include them inline with razor
pages. This is very important because then we can style the svgs with
css.