LiveVideoRenderer.java Performance improvements.
It seems that our parrot colleagues say there is a frame ready
too fast for us to keep up. We are fine with a refresh every 30
seconds. Also if the refresh is too fast and we do not finish
the opengl operations the output on some devices becomes corrupted.
We now make sure we finish the operations of frame rendering.