Gallery: Rewrote gallery front end in Vue.
Most of the same functionality that existed before
but now we also can select multiple files to apply actions
to. Another important improvement is that we no longer
download all the videos for the gallery, but instead load
it dynamically.

Most of the old js and stylesheets that were created before this
commit for the Gallery were removed. There has been a disabling
of the media type selections as well as per timestamp tagging. This
has low priority for now, and more effort for this is not required
when there are other issues to be solved.