vue: Changed project into a MPA style.
By default Vue considers all the projects set up as a single page
thing, but we actually are mixing vue and razor pages so we
want multiple pages. Fortunately Vue allows us that and this commit
is the factoring of the only page we have, the gallery, into it's own
page Vue namespace.