feat(Startup.cs) Persist keys to directory.
Let's try to persist the keys to a directory instead of
keeping it in memory. This way from what i understood
whenever the server is restarted the keys are retrieved
from the local disk instead of being re-generated each
time. This allows for the app user token to be useful
over time