feat(DroHub) Added Account management functionality.
Even though we still have the admin panel it is now
hidden and will eventually be retired to be completely
replaced by the account management page we added in this
commit. Even so, there are different focus for the AdminPanel
and the Account Management functionality now added.

The admin panel is mostly for admin purposes while the new
Account management panel is user facing. It allows for the user
to manage other members of the subscription, including changing
permissions or excluding them, if with appropriate permissions.

It is also possible to invite users with their email address,
whereby the recepient of the email will have it's account activated
upon clicking on the link of the email and confirming it's password.

This commit contains the backend, frontend and tests to the backend.
It also contains some credentials and tests for mail slurp but currently
there is an issue with our account so...