fix(QueryDeviceInfo) Problems with authorization and error report.
There was no way to authorize device information
querying, meaning anybody without authorization could
see other subscription's devices.

Also the reporting for errors, was error prone and there
were scenarios in the application where it got confused
into assuming a device was registered or available when it
was not. Tests were added for the situation.