fix(docker-janus): Use external janus backport.
We needed to update to this version as it seems the
version of janus in vanila ubuntu is defective. We
already had massive issues when we updated, with the
consequence of Firefox not working. Recently even
chrome stopped working with some issues in the
certificate authentication. I just got fed up with it
and upgraded to a newer version. Suddenly firefox
works again