fix(Dashboard): Better connection and env setup.
We were doing some very nasty assumptions, like that there
were only one possible audio remote, when this is not true
when there are more participants in the room.

Also we were not properly joining the rooms and could not recover
when other rooms became available, now we do.

Also it is not really possible to make the rooms permanent
on the backedn due to the fact we use the room setup
on connection from the app to setup connection sessions.
There it goes a big simplification opportunity out of the
window.