fix(parrot-backend): Advise the user if backend connection failed.
Up to now we relied on the fact that if we queryied the
device info then it meant that the telemetry backend was
going to work. While in normal cases this is true we need
to let the user know if something went wrong. Otherwise he
will just see everything working but the platform will not
work.