test(IntegrationTest): Test through nginx rather than dotnet directly.
This way we can guarantee our tests also validate the
infrastructure configuration. Up to this point we directly
bypassed the reverse proxy in nginx and went straight to
the dotnet server. Now we test through the same URI and
paths as the productions. We also test throug https, with
whatever certificate is in the certificates-volume. Due
to high likelyhood of running the tests from localhost we
needed to disable the certificate validation. Sad but needed.

This change on using http infrastructure will also allow human
QA to test video infra through HTTPs and thus not hitting the
limitation most browsers impose on webrtc http.