feat(CreateSubscription): Users can now create subscriptions.
In this commit we add the possibility for the users to choose
a subscription on the landing page. We needed to also add a
privacy policy and terms as this is required for the use of
data gathering.

The subscription plans allowances are now in the appsettings
files.

Small adjustments were done to the subscription elements of the
landing page so that it is more aestetically pleasing.