fix(MJRPostProcessor): Audio sync issues.
There were audio synchronization issues with the live video
recordings due to using the Timespan.Milliseconds instead of
TotalMillisenconds. The former gives the Milliseconds component
of the timespan, while the latter gives the timespan converted
in millisencond units.

Changed the TestAssets to include files multiple audio sources.
Not much of a better test.