refactor(dotnet docker)
Improvements in the order of the build allow for efficient use
of the docker cache, and a much optimized path for when the
web backend or front end change.

Also stopped putting the wwwroot/vue in the docker context
because the contents of that directory should be generated
in the docker build.