Subscription: Added more subscription data and present it.
We now present extra subscription info both on creation
of the subscription as well as on the the AccountManagement page
so that the user can know and communicate their status.

Specifically we present the Price, Allowed Users, Remaining video
minutes and subscription name.

Adjustments made on the CreateSubscription page so that we survive
broken redirects.

The appsettings were updated with the new data.
Initial Admin account has a special subscription name.