MainActivity: Do not make use of NODEVICE.
We no longer need the NODEVICE device as we have proper
status communication on the QueryDeviceInfo. Due to this
hack not being removed, and the logic of QueryDeviceInfo
changing when people were starting the application without
any drone connected they were seeing that the NODEVICE was
not available or they had no permissions. A mess.