AndroidApplicationController: QueryDeviceInfo to require PerformFlightActions claim.
The QueryDeviceInfo is to be only called when from the application
and thus from pilots or higher acting users, that have authorization
to perform flight actions. This is a bigger restriction than the
previous authorization to read the Device information because it
will not authorize the api call if there are no flight minutes left
in the subscription. Probably the QueryDeviceInfo method should be
renamed something more related to pre-flight information query or
something like this, but this would require an API change that would
break tests and application. Technical credit requested...