test: Fixed incorrect validation of invalid token.
We were relying on the unauthorized http status to know if the token
was successfully retrieved or not. We recently changed the token
response to emit nok, and did not update the tests. This led to
tests passing when they should not. Added a test that will guarantee
us we are processing the mobile token properly.