app: Added the concept of lobby.
This is a work in progress where we try to create an intermediate
place between logging in and going to the live flight. Right now
there is not much more functionality besides more informative and
rich situational awareness of the state the application with regards
to the connected devices.

Also includes a DrohubDevice class that tries to abstract the parrot
logic from the app. This allows for an extensible and more interfaced
architecture. Lets see how it turns out.

Heavy use of fragments