TestFixture: Added db and configuration facilities
Due to the deeper intergration of the logic it became possible to read the appsettings
contents and parse it to be available for the test fixture. With this fixture the db context
also became available although not in use in this commit.

The goal of the dbcontext being available is so that we can test some side effects that
are otherwise very hard to measure. Often these side effects are only visible through
visual ways, so it is not really that we are breaking the observability primacy of
our tests.