IntegrationTest: Improved TestChangePermissions
Up to now there was not a test that checked that the claim changes
due to change in permissions were in fact correct. This has been a
bug before and now we have a propper test so it never happens again.

This makes use of the DBContext to verify the side effects. Its the
only way to check it, as the claims values are mostly evaluated in
frontend razor pages.