TestServerFixture: Proper waiting for nginx.
The previous wait mechanism for wating was broken as it
is not guaranteed that the :5000 will be reachble. What
we want is to check if the nginx gateway works. The issue
is that for it to work we need to not check for the ssl
certificate, as the ssl certificate for testing is most
likely self signed for localhost.