feat(app): Notify user through an interface.
Before we were passing android views around to non-android logic
due there being no other way to communicate errors or information
in an abstracted way. Introduced IInfoDisplay interface so that
we pass instances of that interface to classes that might need to
display data to the user. This makes it easier to separate what is
android and what is not, and consequently enable easier testing(no
testing is added in this commit).

InfoDisplayBase base class is created to abstract the information
cycling and lifetime from the actual display. For example SnackBarInfoDisplay
class subclasses InfoDisplayBase to just add Android elements. A
similar but a bit more complicated system was also implemented for
ErrorTextView.