app: Tie the zoom scale with actual value.
Before we initialized the zoom blindly and operated on
it with a made up scale. Now we apply the squeeze and
dilate to the value that the camera zoom reports. We needed
to have a mechanism for not setting the zoom in so small amounts
of scale change because the parrot anafi zoom setting did not
respond to so fine grained changes. Maybe this is not even a
time related unresponsiveness but a magnitude related reponsiveness,
so it could be that accumulating a significant change is more
important than just letting the time pass.