app: Isolate camera and gimbal logic.
The scaling and scrolling logic became dumber and all the clamping
and transformation will now happen in the peripheral. This is
because the peripheral knows better what are the constraints
on camera and gimbal actions.

Try to dissuade the direct access of recorded information of
the peripherals with throwable exceptions on not ready fetch
attempts. On the other hand encourage event driven usage of
peripherals.