app Peripherals: Attempt at abstracting parrot layer into a DROHUB layer.
Lets try to not pass Parrot specific data structures. For that we
no longer expose listeners to Parrot Peripherals but to generic
peripherals. This led to some troubles on the ParrotStreamServer
as the PeerConnectionParamenters required a parrot specific object.
We now generate the ParrotConnectionStreamParameters from the
ParrotStreamServer class. Not ideal as the parrot data structure
is still visible, but soon it will not be.