app: Created a capturer interface that is passed in PeerConnectioParameters.
No longer do we need to pass capturer specific connection parameters because
they should be enclosed in a capturer passed by a CapturerGenerator.

This enables the removal of PeerConnectionParrotStreamParameters as the
capturer generation and it's parameters were moved to the ParrotStreamServer
peripheral class. PeerConnectionParameters package is also gone as there
should not be any reason for more than one PeerConnectionParameters.

Capturer providers should implement IPeripheral.ICapturerProvider. This
allows for an onCapturerAvailable callback that provides an initialized
capturer for a listener to consume.