ParrotPeripheralManager: Remove state as much as possible.
Try to disallow the possibility of having the listener as a mutable.
There was no code needing it so it was easy to do it. Also, Parrot
peripheral come in Ref objects that need to be closed, thus our
ParrotPeripheralManager also needs to close these refs when we
dispose of our Peripherals.