GroundSDKVideoCapturer: Moved StreamServer setup out.
The start and stop stream operations are deeply connected with
the way parrot stream server peripheral works and it's resources.
The capturer should only be concerned in video output not in
peripheral management. We leave that to an Interface called
IParrotStreamServerControll that needs to be implemented by
whatever wants to create the capturer.

Now that we have the state of the stream server managed in the
ParrotStreamServer peripheral class we are even able to fix an
issue where the stream server sink was not closed properly. This
fixed a long running bug where the application crashed when the
drone got disconnected. Its good to not need to have the capturer
managing stream server details.