DroneMicroServiceManager: Mandate explicit action repeat period per service.
Probably due to shortcuts, 5 seconds was deemed a good enough backoff
value to repeat a request for a given telemetry to *all service types.
This is unnacceptable for live telemetry.  Specifically this will lead
to a higher chance of missing events for telemetry. Only pingService and
getLiveVideoState service types are expected to have a long refresh
period as they are not drone but app bound requests, and thus should
have answers on request. These ones can have more relaxed request times.