ParrotMediaStore: Fixed permanent Media Store not available message.
When we made ParrotMediaStore not allow for the normal listeners
we introduced a bug where the uninitialized message of Media store
not available would stay until the first picture was taken. Obviously
this was confusing and reported wrong information. Now ParrotMediaStore
implements the non Parrot specific IPeripheral that restores the onFirstTimeAvailable
callback again, allowing for the removal of the message as soon as
the media store is alive.