ParrotPeripheral: Force the onFirstTimeAvailable to run.
Before we were runnning the onFirstTimeAvailable on the onChange
of the peripheral observer, but this led to some peripherals never
triggering the onChange due there not actually being any change. We
now get() the peripheral handle and force it to be available on demand.
If it is not we throw a runtime error as according to the documentation
this means the peripheral does not exist. If the peripheral exists we
run the onFirstTimeAvailable immediately. Without this change we were
racing to catch an initial event that may have been sent before we were
even set up.