ParrotPeripheralBase: Added a start method.
The start method is needed so that the setting of the
listeners and subclass initialization can happen without
and already running parrot peripheral manager loop. This
could create races and make the state management quite harder.
The other way would be to make the constructor request all the
listeners and set them immediatly but in Java the base class constructor
needs to be run before any other logic which sucks and would make
the issue we are trying to resolve again visible.