Added full stack support for gimbal setting.
Gimbal management in the app is now done through a parrot peripheral just
like the other peripherals. This allowed us to have a common interface
for gimbal setting and notifications compatible with the Thrift RPC interface.

A database table was added to store the GimbalState. Each column has a lot of information,
some of which is usable by a new GimbalSlider in the front end.

Tests setup was modified to also support the gimbal interaction testing.