AdminPanel: Attempt at getting removing DB races.
For some reason there are db races occurring in the tests in the AdminPanel.
I tried to investigate and these changes improve but do not solve the issue.
I think it has to do with the fact that telemetry processes may be ongoing
or winding down while trying to remove the users.