DroHub.Tests: Corrected issues with SignalR telemetry client.
When we changed the uri of drohub to be the https one we
let the signalr one remain http. This was to bypass the hassle
of having to disable the certificate checks for our self signed
certs. The problem is that the cooking passed to the signalr
hub connection did not match the signalr connection and thus
we were not considered authenticated as the user of the cookie.
This led to us never receiving the signalr telemetry.