DroHub.Test: Added patching of docker-compose file for tests.
Up to now it was impossible to run the tests with the unmodified
docker-compose file due to the fact that a certificate is needed
when we are not on drohub.xyz. To still have a master docker-compose
file which is used as the true base for the tests and avoid forking,
every time the tests are run, a patched version of the docker compose
file is generated that is suitable for tests.

To allow for a valid yaml file we needed to set the volumes to something
as the YamlDotNet library did not like yaml nodes without anything set.

Aso disabled the janus service massive assignement of ports, because it kills
docker daemon for some reason